♻️ Extract public/connector.go from connector.go
